feat: Generalize OutPoint struct to reference headers
BREAKING CHANGE: `OutPoint` structure changed
