+++ lib/ckb/types/types.rb
@@ -3,6 +3,7 @@
 require_relative "block_header"
 require_relative "block"
 require_relative "input"
+require_relative "out_point_cell"
 require_relative "out_point"
 require_relative "output"
 require_relative "cell_with_status"
