+++ spec/ckb/types/block_test.rb
@@ -4,7 +4,8 @@ RSpec.describe CKB::Types::Block do
       "transactions": [
         {
           "deps": [
-            {
+            "block_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+            "cell": {
               "tx_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
               "index": 4294967295
             }
@@ -16,8 +17,11 @@ RSpec.describe CKB::Types::Block do
                 "0x0100000000000000"
               ],
               "previous_output": {
-                "tx_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-                "index": 4294967295
+                "block_hash": nil,
+                "cell": {
+                  "tx_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+                  "index": 4294967295
+                }
               },
               "since": "0"
             }
